Handle missing OIDC groups claim
Fixes #6053
Closes #5974
Modifies the behaviour of OpenIDProvider to insert an empty groups list when the groups claim is not present in the token's claims.
This handles the behaviour of some IdPs that omit including empty claims.
This behaviour only gets triggered after the userinfo endpoint is queried in order to not break some existing workflows (access token is missing claims but userinfo endpoint includes those claims).